+++ src/database/index.ts
@@ -1,0 +1,3 @@
+import { Database } from './logic';
+
+export { Database };
