+++ src/database/logic/index.ts
@@ -1,0 +1,3 @@
+import Database from './database';
+
+export { Database };
