+++ src/game-core/entities/business/index.ts
@@ -1,0 +1,13 @@
+// Logic
+import { Business } from './logic';
+
+// Services
+import { BusinessService } from './services';
+
+
+export * from './typings';
+
+export {
+  Business,
+  BusinessService
+};
