+++ src/game-core/entities/business/logic/business.ts
@@ -1,33 +1,15 @@
+// Typings
+import {
+  IBusiness,
+  BusinessIdType,
+  BusinessConfigType
+} from '../typings';
 
-export type BusinessConfigType = {
-  id: BusinessIdType,
-  title: string,
-  price: number,
-  profit: number,
-  upgradePriceMultiplier: number,
-  gainCapitalDurationMs: number
-};
+// Utils
+import { calculateUpgradedPrice } from '../utils/business.utils';
 
-export type BusinessIdType = number;
 
-export interface IBusiness {
-  id: BusinessIdType
-  title: string
-  profit: number
-  level: number
-  price: number
-  gainCapitalDurationMs: number
-  isGainingCapital: boolean
-  upgrade: () => void
-  gainCapital: (callback: (gainedMoney: number) => void) => void
-}
-
-//TODO: move to utils
-const calculateUpgradedPrice = (price: number, upgradePriceMultiplier: number): number => (
-  price + price * upgradePriceMultiplier
-);
-
-export class Business implements IBusiness {
+class Business implements IBusiness {
   private _id:  BusinessIdType;
 
   private _priceMultiplier: number;
@@ -123,3 +105,5 @@ export class Business implements IBusiness {
     this._profit = this.profit + this._initialProfit;
   }
 }
+
+export default Business;
