+++ src/game-core/entities/business/logic/index.ts
@@ -1,0 +1,3 @@
+import Business from './business';
+
+export { Business };
