+++ src/game-core/entities/business/services/index.ts
@@ -1,0 +1,3 @@
+import BusinessService from './business.service';
+
+export { BusinessService };
