+++ src/game-core/entities/business/typings/index.ts
@@ -1,0 +1,2 @@
+
+export * from './business.typings';
