+++ src/game-core/entities/business/utils/business.utils.ts
@@ -1,0 +1,4 @@
+
+export const calculateUpgradedPrice = (price: number, upgradePriceMultiplier: number): number => (
+  price + price * upgradePriceMultiplier
+);
