+++ src/game-core/entities/index.ts
@@ -1,0 +1,8 @@
+import { Player } from './player';
+import { BusinessIdType, BusinessConfigType } from './business';
+
+export {
+  Player,
+  BusinessIdType,
+  BusinessConfigType
+};
