+++ src/game-core/entities/player/index.ts
@@ -1,0 +1,3 @@
+import { Player } from './logic';
+
+export { Player };
