+++ src/game-core/entities/player/logic/index.ts
@@ -1,0 +1,3 @@
+import Player from './player';
+
+export { Player };
