+++ src/game-core/entities/player/typings/index.ts
@@ -1,0 +1,2 @@
+
+export * from './player.typings';
