+++ src/game-core/index.ts
@@ -1,0 +1,3 @@
+import { Player, BusinessIdType, BusinessConfigType } from './entities';
+
+export { Player, BusinessIdType, BusinessConfigType };
