+++ tsconfig.json
@@ -10,7 +10,10 @@
     "target": "es2015",
     "module": "commonjs",
     "moduleResolution": "node",
-    "baseUrl": "./",
+    "baseUrl": ".",
+    "paths": {
+      "@src/*": ["src/*"]
+    },
     "strict": true,
     "outDir": "./dist",
     "resolveJsonModule": true
